+++ src/main/java/chess/controller/ExceptionController.java
@@ -1,0 +1,24 @@
+package chess.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import java.sql.SQLException;
+
+import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
+
+@ControllerAdvice
+public class ExceptionController {
+
+    @ExceptionHandler({IllegalArgumentException.class})
+    public ResponseEntity<String> handleIllegalArgumentException() {
+        return ResponseEntity.badRequest().body("unavailable");
+    }
+
+    @ExceptionHandler({SQLException.class})
+    public ResponseEntity<String> handleSQLException() {
+        return ResponseEntity.status(INTERNAL_SERVER_ERROR).body("fail");
+    }
+    
+}
