+++ src/main/java/chess/controller/SpringChessController.java
@@ -34,14 +34,4 @@ public class SpringChessController {
         return springChessService.move(moveRequestDto);
     }
 
-    @ExceptionHandler({IllegalArgumentException.class})
-    public ResponseEntity<String> handleIllegalArgumentException() {
-        return ResponseEntity.badRequest().body("unavailable");
-    }
-
-    @ExceptionHandler({SQLException.class})
-    public ResponseEntity<String> handleSQLException() {
-        return ResponseEntity.status(INTERNAL_SERVER_ERROR).body("fail");
-    }
-
 }
